restored the software renderering (non-opengl) waveforms, deprecated all qt(old qglwidget based), gl and glsl waveform widgets
