+++ src/waveform/renderers/glwaveformrenderbackground.h
