+++ src/waveform/widgets/deprecated/qthsvwaveformwidget.h
