+++ src/waveform/widgets/deprecated/qtrgbwaveformwidget.cpp
@@ -1,47 +1,53 @@
-#include "waveform/widgets/qtvsynctestwidget.h"
+#include "waveform/widgets/qtrgbwaveformwidget.h"
 
 #include <QPainter>
 #include <QtDebug>
 
-#include "moc_qtvsynctestwidget.cpp"
-#include "util/performancetimer.h"
+#include "moc_qtrgbwaveformwidget.cpp"
 #include "waveform/renderers/glwaveformrenderbackground.h"
-#include "waveform/renderers/qtvsynctestrenderer.h"
 #include "waveform/renderers/waveformrenderbeat.h"
 #include "waveform/renderers/waveformrendererendoftrack.h"
 #include "waveform/renderers/waveformrendererpreroll.h"
+#include "waveform/renderers/waveformrendererrgb.h"
 #include "waveform/renderers/waveformrendermark.h"
 #include "waveform/renderers/waveformrendermarkrange.h"
 #include "waveform/renderers/waveformwidgetrenderer.h"
 
-QtVSyncTestWidget::QtVSyncTestWidget(const QString& group, QWidget* parent)
+QtRGBWaveformWidget::QtRGBWaveformWidget(const QString& group, QWidget* parent)
         : GLWaveformWidgetAbstract(group, parent) {
-    addRenderer<QtVSyncTestRenderer>();
+    addRenderer<GLWaveformRenderBackground>();
+    addRenderer<WaveformRendererEndOfTrack>();
+    addRenderer<WaveformRendererPreroll>();
+    addRenderer<WaveformRenderMarkRange>();
+    addRenderer<WaveformRendererRGB>();
+    addRenderer<WaveformRenderBeat>();
+    addRenderer<WaveformRenderMark>();
+
     m_initSuccess = init();
 }
 
-QtVSyncTestWidget::~QtVSyncTestWidget() {
+QtRGBWaveformWidget::~QtRGBWaveformWidget() {
 }
 
-void QtVSyncTestWidget::castToQWidget() {
+void QtRGBWaveformWidget::castToQWidget() {
     m_widget = this;
 }
 
-void QtVSyncTestWidget::paintEvent(QPaintEvent* event) {
+void QtRGBWaveformWidget::paintEvent(QPaintEvent* event) {
     Q_UNUSED(event);
 }
 
-mixxx::Duration QtVSyncTestWidget::render() {
+mixxx::Duration QtRGBWaveformWidget::render() {
     PerformanceTimer timer;
     mixxx::Duration t1;
-    //mixxx::Duration t2, t3;
+    // mixxx::Duration t2, t3;
     timer.start();
     // QPainter makes QGLContext::currentContext() == context()
     // this may delayed until previous buffer swap finished
     QPainter painter(paintDevice());
     t1 = timer.restart();
     draw(&painter, nullptr);
-    //t2 = timer.restart();
-    //qDebug() << "GLVSyncTestWidget "<< t1 << t2;
+    // t2 = timer.restart();
+    // qDebug() << "GLVSyncTestWidget "<< t1 << t2;
     return t1; // return timer for painter setup
 }
