+++ src/waveform/widgets/glwaveformwidgetabstract.h
